feat: hide internal functionalities from side navigation (incl. TypeScripting)
